% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedians.R
\docType{methods}
\name{rowMedians,data.frame-method}
\alias{rowMedians,data.frame-method}
\title{Toy implementation of rowMedians for data.frame input}
\usage{
\S4method{rowMedians}{data.frame}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x))
}
\arguments{
\item{x}{An NxK \code{\link[base]{data.frame}} object.}

\item{rows, cols}{Not yet implemented.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{Not yet implemented.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Toy implementation of rowMedians for data.frame input.
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMedians}()}
}
}
